drop check for terminals
From the original author:

  "Note that this only enables bracketed paste mode when the $TERM value
  starts with "xterm".. you could omit the test altogether, since it
  isn’t strictly necessary—it’s just trying to be careful not to do
  something that might be incompatible with some other terminal type."
  (http://stackoverflow.com/questions/5585129/pasting-code-into-terminal-window-into-vim-on-mac-os-x/7053522#7053522)

Since we don't understand yet if this is a problem, I recommend dropping
the check and seeing if it comes back to bite us. Should be safe since
vim has undo.

(My machine has term set to 'linux', and I'm reluctant to keep adding
clauses to that check.)